+++ src/migrations/1.1.x/m150714_000001_alter_audit_data.php
@@ -1,0 +1,15 @@
+<?php
+
+use bedezign\yii2\audit\components\Migration;
+use yii\db\Schema;
+
+class m150714_000001_alter_audit_data extends Migration
+{
+    const TABLE = '{{%audit_data}}';
+
+    public function up()
+    {
+        $this->addColumn(self::TABLE, 'created', Schema::TYPE_DATETIME);
+    }
+
+}