+++ src/models/AuditEntrySearch.php
@@ -3,6 +3,7 @@
 namespace bedezign\yii2\audit\models;
 
 use bedezign\yii2\audit\Audit;
+use bedezign\yii2\audit\components\DbHelper;
 use yii\base\Model;
 use yii\data\ActiveDataProvider;
 use yii\db\ActiveQuery;
@@ -40,6 +41,7 @@ class AuditEntrySearch extends AuditEntry
     public function search($params)
     {
         $query = AuditEntry::find();
+        $query->select($this->safeAttributes());
 
         $dataProvider = new ActiveDataProvider([
             'query' => $query,
@@ -64,7 +66,7 @@ class AuditEntrySearch extends AuditEntry
         $query->andFilterWhere(['ajax' => $this->ajax]);
         $query->andFilterWhere(['duration' => $this->duration]);
         $query->andFilterWhere(['memory_max' => $this->memory_max]);
-        $query->andFilterWhere(['like', 'created', $this->created]);
+        $query->andFilterWhere(['like', DbHelper::convertIfNeeded(AuditEntry::class, 'created', 'text'), $this->created]);
         $query->with(['linkedErrors', 'javascripts']);
 
         return $dataProvider;
@@ -102,11 +104,9 @@ class AuditEntrySearch extends AuditEntry
      */
     protected function filterUserId($userId, $query)
     {
-        if (strlen($userId)) {
+        if (!empty($userId)) {
             if (!is_numeric($userId) && $callback = Audit::getInstance()->userFilterCallback) {
                 $userId = call_user_func($callback, $userId);
-            } else {
-                $userId = intval($userId) ?: 0;
             }
             $query->andWhere(['user_id' => $userId]);
         }